end developer

+++ app/Http/Controllers/LikeController.php
@@ -23,6 +23,7 @@ class LikeController extends Controller
             'track_id' => 'required|numeric'
         ];
         $validator = Validator::make($request->input(), $rules);
+
         if ($validator->fails()) {
             return response()->json([
                 'status' => false,
@@ -31,6 +32,13 @@ class LikeController extends Controller
         }
         $like = new Like($request->input());
         $like->save();
+        
+        // Incrementar el contador de likes en el track correspondiente
+        $track = Track::find($request->input('track_id'));
+        if ($track) {
+            $track->increment('likes');
+        }
+
         return response()->json([
             'status' => true,
             'message' => 'Like GUARDADO correctamente',
@@ -45,7 +53,22 @@ class LikeController extends Controller
 
     public function destroy(Like $like)
     {
+        // Obtener el track asociado al like
+        $track = Track::find($like->track_id);
+
+        if (!$track) {
+            return response()->json([
+                'status' => false,
+                'message' => 'No se encontró el track asociado al like'
+            ], 404);
+        }
+
+        // Eliminar el like
         $like->delete();
+
+        // Decrementar el contador de likes en el track
+        $track->decrement('likes');
+
         return response()->json([
             'status' => true,
             'message' => 'Like BORRADO correctamente'
@@ -81,7 +104,8 @@ class LikeController extends Controller
         // Obtener los tracks que tienen likes del usuario
         $tracks = Track::whereHas('likes', function ($query) use ($user_id) {
             $query->where('user_id', $user_id);
-        })->get();
+        })->orderBy('updated_at', 'desc')
+        ->paginate(8);
 
         // Devolver la respuesta
         return response()->json(['tracks' => $tracks]);
